metal/pyproject.toml
[project]
name = "metal"
version = "0.1.0"
description = "The backend server for the Compass Journal."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "openai",
    "flask",
    "flask-cors",
    "werkzeug",
    "requests",
    "firebase-admin"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["metal"]